YouTube Downloader App using pytube-fix
This program will download a youtube video at the highest resolution, or you can select audio file only.

I found myself wanting this very tool, but with less features. I wanted to make it accessible to people who don't exactly know python. Please utilize the README for instructions on how to use and how the program operates.

The only files you need are the .exe and the .ico if you want the icon to work in the taskbar. Initially the file will expect the ico to be "next to", in the same folder or directory as the .exe file, but you can change it in the .json file that will be generated on first launch. This file must remain in the same folder/directory as the .exe file in order to retain your last save. If the program does not find that file in it's directory then it will generate a new one with default parameters.